+++ app/src/main/java/com/capstone/milkyway/viewmodel/DonationViewModel.kt
@@ -26,7 +26,7 @@ class DonationViewModel : ViewModel() {
         userId: String,
         name: String,
         age: Int,
-        phone: Int,
+        phone: Long,
         religion: String,
         healthCondition: String,
         isSmoking: String,
